refactor: 精简设置界面文件体积
