+++ src/renderer/components/states/maxExpand/components/SettingsTab.tsx
@@ -27,7 +27,6 @@
 import { useState, useRef, useEffect, useMemo } from 'react';
 import type { KeyboardEvent, ReactElement, ReactNode } from 'react';
 import useIslandStore from '../../../../store/slices';
-import avatarImg from '../../../../assets/avatar/T.jpg';
 import type { OverviewWidgetType, OverviewLayoutConfig } from '../../expand/components/OverviewTab';
 import { OVERVIEW_WIDGET_OPTIONS } from '../../expand/components/OverviewTab';
 import { SvgIcon } from '../../../../utils/SvgIcon';
@@ -44,6 +43,38 @@ import {
   loadWeatherLocationConfig,
   saveWeatherLocationConfig,
 } from '../../../../store/utils/storage';
+import {
+  LYRICS_SOURCE_OPTIONS,
+  WEATHER_PROVIDER_OPTIONS,
+  WEATHER_LOCATION_PRIORITY_OPTIONS,
+  SETTINGS_TABS,
+  SETTINGS_TAB_LABELS,
+  NETWORK_TIMEOUT_OPTIONS,
+  LAYOUT_STORE_KEY,
+  DEFAULT_LAYOUT,
+  APP_SETTINGS_PAGES,
+  WEATHER_SETTINGS_PAGES,
+  WEATHER_SETTINGS_PAGE_LABELS,
+  MUSIC_SETTINGS_PAGES,
+  MUSIC_SETTINGS_PAGE_LABELS,
+  NAV_CARDS,
+  DEFAULT_NAV_ORDER,
+  NAV_CARDS_MAP,
+  type SettingsSidebarTabKey,
+  type AppSettingsPageKey,
+  type WeatherSettingsPageKey,
+  type MusicSettingsPageKey,
+  type NavCardDef,
+} from './setting/utils/settingsConfig';
+import { UpdateSettingsSection } from './setting/components/update/UpdateSettingsSection';
+import { IndexSettingsSection } from './setting/components/index/IndexSettingsSection';
+import { AppSettingsSection } from './setting/components/app/AppSettingsSection';
+import { NetworkSettingsSection } from './setting/components/network/NetworkSettingsSection';
+import { WeatherSettingsSection } from './setting/components/weather/WeatherSettingsSection';
+import { ShortcutSettingsSection } from './setting/components/shortcut/ShortcutSettingsSection';
+import { MusicSettingsSection } from './setting/components/music/MusicSettingsSection';
+import { AiSettingsSection } from './setting/components/ai/AiSettingsSection';
+import { AboutSettingsSection } from './setting/components/about/AboutSettingsSection';
 
 import { resolveDistrictLocationByKeyword } from '../../../../api/adcodeApi';
 
@@ -338,145 +369,6 @@ function OverviewPreview({ layoutConfig }: { layoutConfig: OverviewLayoutConfig 
  * 设置 Tab
  * @description 最大展开模式下的设置面板
  */
-/** 歌词源选项 */
-const LYRICS_SOURCE_OPTIONS = [
-  { value: 'auto', label: '自动（跟随播放器）' },
-  { value: 'netease-only', label: '仅网易云' },
-  { value: 'qqmusic-only', label: '仅 QQ音乐' },
-  { value: 'kugou-only', label: '仅酷狗' },
-  { value: 'sodamusic-only', label: '仅汽水音乐' },
-  { value: 'lrclib-only', label: '仅 LRCLIB' },
-];
-
-const WEATHER_PROVIDER_OPTIONS: Array<{ value: WeatherProvider; label: string }> = [
-  { value: 'open-meteo', label: 'Open-Meteo 优先' },
-  { value: 'uapi', label: 'UAPI 优先' },
-];
-
-const WEATHER_LOCATION_PRIORITY_OPTIONS: Array<{ value: WeatherLocationPriority; label: string }> = [
-  { value: 'ip', label: 'IP 定位优先' },
-  { value: 'custom', label: '自定义位置优先' },
-];
-
-/** 设置页侧边栏 Tab 顺序 */
-const SETTINGS_TABS: ('index' | 'app' | 'network' | 'weather' | 'music' | 'ai' | 'shortcut' | 'update' | 'about')[] = ['index', 'app', 'network', 'weather', 'music', 'ai', 'shortcut', 'update', 'about'];
-type SettingsSidebarTabKey = (typeof SETTINGS_TABS)[number];
-type AppSettingsPageKey = 'layout-preview' | 'hide-process-list' | 'position' | 'theme' | 'behavior' | 'autostart';
-type WeatherSettingsPageKey = 'location' | 'provider';
-type MusicSettingsPageKey = 'whitelist' | 'lyrics' | 'smtc';
-type MusicNavCardKey = 'music-whitelist' | 'music-lyrics' | 'music-smtc';
-type SettingsTabLabelKey = SettingsSidebarTabKey | AppSettingsPageKey | MusicNavCardKey;
-
-const SETTINGS_TAB_LABELS: Record<SettingsTabLabelKey, string> = {
-  index: '快速导航',
-  app: '软件设置',
-  'layout-preview': '布局预览',
-  'hide-process-list': '隐藏进程管理',
-  position: '位置校准',
-  theme: '主题外观',
-  behavior: '交互行为',
-  autostart: '实用工具',
-  network: '网络配置',
-  weather: '天气配置',
-  music: '歌曲设置',
-  'music-whitelist': '播放器白名单',
-  'music-lyrics': '歌词源',
-  'music-smtc': 'SMTC',
-  ai: 'AI Agent',
-  shortcut: '快捷键',
-  update: '更新设置',
-  about: '关于软件',
-};
-const SETTINGS_TAB_DESCRIPTIONS: Record<Exclude<SettingsTabLabelKey, 'index'>, string> = {
-  app: '布局预览与隐藏进程规则配置',
-  'layout-preview': '进入布局预览并调整左右控件展示。',
-  'hide-process-list': '管理隐藏进程名单与自动隐藏规则。',
-  position: '动态调整灵动岛位置并保存',
-  theme: '切换深色、浅色或跟随系统主题。',
-  behavior: '配置鼠标移开后是否自动收回。',
-  autostart: '应用控制、日志与开机启动配置。',
-  network: '请求超时与网络行为设置',
-  weather: '天气接口优先级设置',
-  music: '播放器白名单与歌词来源',
-  'music-whitelist': '配置允许接入灵动岛的播放器。',
-  'music-lyrics': '选择歌词来源与显示模式。',
-  'music-smtc': '系统媒体传输控制相关配置。',
-  ai: 'AI 服务与 Prompt 配置',
-  shortcut: '隐藏、关闭、截图快捷键',
-  update: '检查与下载软件更新',
-  about: '版本信息与项目链接',
-};
-const SETTINGS_TAB_ICONS: Partial<Record<SettingsTabLabelKey, string>> = {
-  'layout-preview': SvgIcon.LAYOUT,
-  'hide-process-list': SvgIcon.TASK_MANAGER,
-  position: SvgIcon.MOVE,
-  network: SvgIcon.NETWORK,
-  weather: SvgIcon.WEATHER,
-  music: SvgIcon.LRC,
-  'music-whitelist': SvgIcon.MUSIC,
-  'music-lyrics': SvgIcon.LRC,
-  'music-smtc': SvgIcon.SMTC,
-  ai: SvgIcon.AI,
-  shortcut: SvgIcon.SHORTCUT_KEY,
-  update: SvgIcon.REVERT,
-  about: SvgIcon.ABOUT,
-  theme: SvgIcon.THEME,
-  behavior: SvgIcon.INTERACTION,
-  autostart: SvgIcon.CONTINUE,
-};
-
-const NETWORK_TIMEOUT_OPTIONS = [
-  { label: '5 秒', value: 5000 },
-  { label: '10 秒（默认）', value: 10000 },
-  { label: '15 秒', value: 15000 },
-  { label: '20 秒', value: 20000 },
-  { label: '30 秒', value: 30000 },
-];
-
-const LAYOUT_STORE_KEY = 'overview-layout';
-const DEFAULT_LAYOUT: OverviewLayoutConfig = { left: 'shortcuts', right: 'todo' };
-const APP_SETTINGS_PAGES: AppSettingsPageKey[] = ['layout-preview', 'hide-process-list', 'position', 'theme', 'behavior', 'autostart'];
-const WEATHER_SETTINGS_PAGES: WeatherSettingsPageKey[] = ['location', 'provider'];
-const WEATHER_SETTINGS_PAGE_LABELS: Record<WeatherSettingsPageKey, string> = {
-  location: '定位配置',
-  provider: '接口配置',
-};
-const MUSIC_SETTINGS_PAGES: MusicSettingsPageKey[] = ['whitelist', 'lyrics', 'smtc'];
-const MUSIC_SETTINGS_PAGE_LABELS: Record<MusicSettingsPageKey, string> = {
-  whitelist: '白名单',
-  lyrics: '歌词源',
-  smtc: 'SMTC',
-};
-
-interface NavCardDef {
-  id: string;
-  label: string;
-  desc: string;
-  icon?: string;
-  tab: SettingsSidebarTabKey;
-  appPage?: AppSettingsPageKey;
-  musicPage?: MusicSettingsPageKey;
-}
-
-const NAV_CARDS: NavCardDef[] = [
-  { id: 'layout-preview', label: SETTINGS_TAB_LABELS['layout-preview'], desc: SETTINGS_TAB_DESCRIPTIONS['layout-preview'], icon: SETTINGS_TAB_ICONS['layout-preview'], tab: 'app', appPage: 'layout-preview' },
-  { id: 'hide-process-list', label: SETTINGS_TAB_LABELS['hide-process-list'], desc: SETTINGS_TAB_DESCRIPTIONS['hide-process-list'], icon: SETTINGS_TAB_ICONS['hide-process-list'], tab: 'app', appPage: 'hide-process-list' },
-  { id: 'position', label: SETTINGS_TAB_LABELS.position, desc: SETTINGS_TAB_DESCRIPTIONS.position, icon: SETTINGS_TAB_ICONS.position, tab: 'app', appPage: 'position' },
-  { id: 'theme', label: SETTINGS_TAB_LABELS.theme, desc: SETTINGS_TAB_DESCRIPTIONS.theme, icon: SETTINGS_TAB_ICONS.theme, tab: 'app', appPage: 'theme' },
-  { id: 'behavior', label: SETTINGS_TAB_LABELS.behavior, desc: SETTINGS_TAB_DESCRIPTIONS.behavior, icon: SETTINGS_TAB_ICONS.behavior, tab: 'app', appPage: 'behavior' },
-  { id: 'autostart', label: SETTINGS_TAB_LABELS.autostart, desc: SETTINGS_TAB_DESCRIPTIONS.autostart, icon: SETTINGS_TAB_ICONS.autostart, tab: 'app', appPage: 'autostart' },
-  { id: 'network', label: SETTINGS_TAB_LABELS.network, desc: SETTINGS_TAB_DESCRIPTIONS.network, icon: SETTINGS_TAB_ICONS.network, tab: 'network' },
-  { id: 'weather', label: SETTINGS_TAB_LABELS.weather, desc: SETTINGS_TAB_DESCRIPTIONS.weather, icon: SETTINGS_TAB_ICONS.weather, tab: 'weather' },
-  { id: 'ai', label: SETTINGS_TAB_LABELS.ai, desc: SETTINGS_TAB_DESCRIPTIONS.ai, icon: SETTINGS_TAB_ICONS.ai, tab: 'ai' },
-  { id: 'shortcut', label: SETTINGS_TAB_LABELS.shortcut, desc: SETTINGS_TAB_DESCRIPTIONS.shortcut, icon: SETTINGS_TAB_ICONS.shortcut, tab: 'shortcut' },
-  { id: 'about', label: SETTINGS_TAB_LABELS.about, desc: SETTINGS_TAB_DESCRIPTIONS.about, icon: SETTINGS_TAB_ICONS.about, tab: 'about' },
-  { id: 'music-whitelist', label: SETTINGS_TAB_LABELS['music-whitelist'], desc: SETTINGS_TAB_DESCRIPTIONS['music-whitelist'], icon: SETTINGS_TAB_ICONS['music-whitelist'], tab: 'music', musicPage: 'whitelist' },
-  { id: 'music-lyrics', label: SETTINGS_TAB_LABELS['music-lyrics'], desc: SETTINGS_TAB_DESCRIPTIONS['music-lyrics'], icon: SETTINGS_TAB_ICONS['music-lyrics'], tab: 'music', musicPage: 'lyrics' },
-  { id: 'music-smtc', label: SETTINGS_TAB_LABELS['music-smtc'], desc: SETTINGS_TAB_DESCRIPTIONS['music-smtc'], icon: SETTINGS_TAB_ICONS['music-smtc'], tab: 'music', musicPage: 'smtc' },
-];
-
-const DEFAULT_NAV_ORDER: string[] = NAV_CARDS.map((c) => c.id);
-const NAV_CARDS_MAP = new Map(NAV_CARDS.map((c) => [c.id, c]));
 
 interface RunningProcessItem {
   name: string;
@@ -1490,1362 +1382,210 @@ export function SettingsTab(): ReactElement {
 
         <div className="max-expand-settings-panel">
           {activeTab === 'index' && (
-            <div className="max-expand-settings-section settings-index-section">
-              <div className="settings-index-header">
-                <div className="max-expand-settings-title">
-                  快速导航
-                  <button
-                    className="settings-nav-edit-btn"
-                    type="button"
-                    onClick={resetNavConfig}
-                  >
-                    恢复默认
-                  </button>
-                  <button
-                    className={`settings-nav-edit-btn ${navEditMode ? 'active' : ''}`}
-                    type="button"
-                    onClick={() => {
-                      if (navEditMode) {
-                        persistNavConfig(navOrder, hiddenNavOrder);
-                      }
-                      setNavEditMode(!navEditMode);
-                    }}
-                  >
-                    {navEditMode ? '完成' : '编辑'}
-                  </button>
-                </div>
-                <div className="settings-music-hint settings-index-hint">
-                  {navEditMode ? '拖拽卡片可调整排列顺序，点击「完成」保存。' : '点击卡片可快速跳转到对应配置页。'}
-                </div>
-              </div>
-              <div className="settings-index-cards" aria-label="设置快速导航">
-                {visibleCards.map((card, idx) => (
-                  navEditMode ? (
-                    <div
-                      key={card.id}
-                      className={`settings-index-card editing${dragOverIdx === idx ? ' drag-over' : ''}`}
-                      draggable
-                      onDragStart={(e) => {
-                        dragIdxRef.current = idx;
-                        e.dataTransfer.effectAllowed = 'move';
-                      }}
-                      onDragOver={(e) => {
-                        e.preventDefault();
-                        setDragOverIdx(idx);
-                      }}
-                      onDragLeave={() => setDragOverIdx(null)}
-                      onDrop={(e) => {
-                        e.preventDefault();
-                        setDragOverIdx(null);
-                        const from = dragIdxRef.current;
-                        if (from === null || from === idx) return;
-                        const newOrder = visibleCards.map((c) => c.id);
-                        const [moved] = newOrder.splice(from, 1);
-                        newOrder.splice(idx, 0, moved);
-                        setNavOrder(newOrder);
-                      }}
-                      onDragEnd={() => {
-                        dragIdxRef.current = null;
-                        setDragOverIdx(null);
-                      }}
-                    >
-                      <span className="settings-index-card-drag-handle">⠿</span>
-                      <button
-                        className="settings-index-card-remove"
-                        type="button"
-                        onClick={() => {
-                          const nextVisible = navOrder.filter((id) => id !== card.id);
-                          const nextHidden = hiddenNavOrder.includes(card.id) ? hiddenNavOrder : [...hiddenNavOrder, card.id];
-                          setNavOrder(nextVisible);
-                          setHiddenNavOrder(nextHidden);
-                        }}
-                        aria-label={`删除 ${card.label}`}
-                      >
-                        −
-                      </button>
-                      <span className="settings-index-card-title">{card.label}</span>
-                      <span className="settings-index-card-desc">{card.desc}</span>
-                      {card.icon && (
-                        <img className="settings-index-card-layout-icon" src={card.icon} alt="" aria-hidden="true" />
-                      )}
-                    </div>
-                  ) : (
-                    <button
-                      key={card.id}
-                      className="settings-index-card"
-                      type="button"
-                      onClick={() => {
-                        if (card.appPage) {
-                          setAppSettingsPage(card.appPage);
-                          setActiveTab('app');
-                        } else if (card.musicPage) {
-                          setMusicSettingsPage(card.musicPage);
-                          setActiveTab('music');
-                        } else {
-                          setActiveTab(card.tab);
-                        }
-                      }}
-                    >
-                      <span className="settings-index-card-title">{card.label}</span>
-                      <span className="settings-index-card-desc">{card.desc}</span>
-                      {card.icon && (
-                        <img className="settings-index-card-layout-icon" src={card.icon} alt="" aria-hidden="true" />
-                      )}
-                    </button>
-                  )
-                ))}
-              </div>
-              {navEditMode && (
-                <div className="settings-nav-add-panel" aria-label="可添加导航卡片">
-                  <div className="settings-music-label">可添加卡片</div>
-                  {hiddenCards.length === 0 ? (
-                    <div className="settings-music-hint">当前没有可添加的卡片</div>
-                  ) : (
-                    <div className="settings-nav-add-list">
-                      {hiddenCards.map((card) => (
-                        <button
-                          key={card.id}
-                          className="settings-nav-add-item"
-                          type="button"
-                          onClick={() => {
-                            const nextVisible = navOrder.includes(card.id) ? navOrder : [...navOrder, card.id];
-                            const nextHidden = hiddenNavOrder.filter((id) => id !== card.id);
-                            setNavOrder(nextVisible);
-                            setHiddenNavOrder(nextHidden);
-                          }}
-                        >
-                          <span>{card.label}</span>
-                          <span className="settings-nav-add-plus">+</span>
-                        </button>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
+            <IndexSettingsSection
+              visibleCards={visibleCards}
+              hiddenCards={hiddenCards}
+              navEditMode={navEditMode}
+              dragOverIdx={dragOverIdx}
+              navOrder={navOrder}
+              hiddenNavOrder={hiddenNavOrder}
+              dragIdxRef={dragIdxRef}
+              setDragOverIdx={setDragOverIdx}
+              setNavOrder={setNavOrder}
+              setHiddenNavOrder={setHiddenNavOrder}
+              setNavEditMode={setNavEditMode}
+              resetNavConfig={resetNavConfig}
+              persistNavConfig={persistNavConfig}
+              setAppSettingsPage={setAppSettingsPage}
+              setMusicSettingsPage={setMusicSettingsPage}
+              setActiveTab={setActiveTab}
+            />
           )}
 
           {activeTab === 'app' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title settings-app-title-line">
-                <span>软件设置</span>
-                <span className="settings-app-title-sub">- {currentAppSettingsPageLabel}</span>
-              </div>
-
-              <div className="settings-app-pages-layout">
-                <div className="settings-app-page-main">
-                  {appSettingsPage === 'layout-preview' && (
-                    <div className="settings-island-preview-section">
-                      <div className="settings-island-preview-label">总览布局预览</div>
-                      <div className="settings-island-preview-wrap">
-                        <div className="settings-island-shell" key={`${layoutConfig.left}-${layoutConfig.right}`}>
-                          <OverviewPreview layoutConfig={layoutConfig} />
-                        </div>
-                      </div>
-
-                      <div className="settings-layout-controls">
-                        <div className="settings-layout-control">
-                          <span className="settings-layout-control-label">左侧控件</span>
-                          <div className="settings-layout-options">
-                            {OVERVIEW_WIDGET_OPTIONS.map(opt => (
-                              <button
-                                key={opt.value}
-                                className={`settings-layout-btn ${layoutConfig.left === opt.value ? 'active' : ''}`}
-                                type="button"
-                                onClick={() => updateLayout('left', opt.value)}
-                              >
-                                {opt.label}
-                              </button>
-                            ))}
-                          </div>
-                        </div>
-                        <div className="settings-layout-control">
-                          <span className="settings-layout-control-label">右侧控件</span>
-                          <div className="settings-layout-options">
-                            {OVERVIEW_WIDGET_OPTIONS.map(opt => (
-                              <button
-                                key={opt.value}
-                                className={`settings-layout-btn ${layoutConfig.right === opt.value ? 'active' : ''}`}
-                                type="button"
-                                onClick={() => updateLayout('right', opt.value)}
-                              >
-                                {opt.label}
-                              </button>
-                            ))}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {appSettingsPage === 'hide-process-list' && (
-                    <div className="settings-hide-processes">
-                      <div className="settings-music-hint">识别到以下进程运行时，将立即隐藏灵动岛。新增进程下次重启生效，删除进程立即生效。</div>
-                      <div className="settings-hide-process-toolbar">
-                        <input
-                          className="settings-whitelist-input"
-                          type="text"
-                          placeholder="搜索进程名（如 WeChat.exe）"
-                          value={hideProcessFilter}
-                          onChange={(e) => setHideProcessFilter(e.target.value)}
-                        />
-                        <button
-                          className="settings-whitelist-add-btn"
-                          type="button"
-                          onClick={() => {
-                            refreshRunningProcesses().catch(() => {});
-                          }}
-                          disabled={hideProcessLoading}
-                        >
-                          {hideProcessLoading ? '刷新中…' : '刷新进程'}
-                        </button>
-                      </div>
-
-                      <div className="settings-hide-selected">
-                        {hideProcessList.length === 0 ? (
-                          <span className="settings-hide-selected-empty">暂无隐藏进程</span>
-                        ) : hideProcessList.map((name) => (
-                          <button
-                            key={name}
-                            className="settings-hide-selected-item"
-                            type="button"
-                            onClick={() => toggleHideProcess(name)}
-                            title="移除该进程"
-                          >
-                            {name} ×
-                          </button>
-                        ))}
-                      </div>
-
-                      <div className="settings-hide-process-list">
-                        {runningProcesses
-                          .filter((process) => process.name.toLowerCase().includes(hideProcessKeyword))
-                          .map((process) => {
-                            const name = process.name;
-                            const selected = hideProcessList.some((item) => item.trim().toLowerCase() === name.trim().toLowerCase());
-                            const fallbackText = name.charAt(0).toUpperCase();
-                            return (
-                              <button
-                                key={name}
-                                className={`settings-hide-process-item ${selected ? 'active' : ''}`}
-                                type="button"
-                                onClick={() => toggleHideProcess(name)}
-                              >
-                                <span className={`settings-hide-process-check ${selected ? 'active' : ''}`}>{selected ? '✓' : ''}</span>
-                                <span className="settings-hide-process-icon" aria-hidden="true">
-                                  {process.iconDataUrl ? (
-                                    <img src={process.iconDataUrl} alt="" />
-                                  ) : (
-                                    <span>{fallbackText || '•'}</span>
-                                  )}
-                                </span>
-                                <span className="settings-hide-process-name">{name}</span>
-                              </button>
-                            );
-                          })}
-                      </div>
-                    </div>
-                  )}
-
-                  {appSettingsPage === 'position' && (
-                    <div className="max-expand-settings-section">
-                      <div className="settings-music-section">
-                        <div className="settings-music-label">灵动岛位置偏移</div>
-                        <div className="settings-music-hint">调整后立即生效并自动保存，重启后会按该位置校准。</div>
-
-                        <div className="settings-hotkey-row">
-                          <button className="settings-hotkey-btn" type="button" onClick={() => applyIslandPositionOffset(islandPositionOffset.x - 10, islandPositionOffset.y)}>左移 10</button>
-                          <button className="settings-hotkey-btn" type="button" onClick={() => applyIslandPositionOffset(islandPositionOffset.x + 10, islandPositionOffset.y)}>右移 10</button>
-                          <button className="settings-hotkey-btn" type="button" onClick={() => applyIslandPositionOffset(islandPositionOffset.x, islandPositionOffset.y - 10)}>上移 10</button>
-                          <button className="settings-hotkey-btn" type="button" onClick={() => applyIslandPositionOffset(islandPositionOffset.x, islandPositionOffset.y + 10)}>下移 10</button>
-                        </div>
-
-                        <div className="settings-hotkey-row">
-                          <label className="settings-field" style={{ flex: 1 }}>
-                            <span className="settings-field-label">水平偏移 X（px）</span>
-                            <input
-                              className="settings-field-input"
-                              type="number"
-                              min={-2000}
-                              max={2000}
-                              value={islandPositionInput.x}
-                              onChange={(e) => {
-                                setIslandPositionInput((prev) => ({ ...prev, x: e.target.value }));
-                              }}
-                              onKeyDown={(e) => {
-                                if (e.key === 'Enter') {
-                                  e.preventDefault();
-                                  applyIslandPositionInput();
-                                }
-                              }}
-                            />
-                          </label>
-                          <label className="settings-field" style={{ flex: 1 }}>
-                            <span className="settings-field-label">垂直偏移 Y（px）</span>
-                            <input
-                              className="settings-field-input"
-                              type="number"
-                              min={-1200}
-                              max={1200}
-                              value={islandPositionInput.y}
-                              onChange={(e) => {
-                                setIslandPositionInput((prev) => ({ ...prev, y: e.target.value }));
-                              }}
-                              onKeyDown={(e) => {
-                                if (e.key === 'Enter') {
-                                  e.preventDefault();
-                                  applyIslandPositionInput();
-                                }
-                              }}
-                            />
-                          </label>
-                        </div>
-
-                        <div className="settings-hotkey-row">
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={applyIslandPositionInput}
-                            disabled={!islandPositionInputChanged}
-                          >
-                            应用
-                          </button>
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={cancelIslandPositionInput}
-                            disabled={!islandPositionInputChanged}
-                          >
-                            取消
-                          </button>
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={() => applyIslandPositionOffset(0, 0)}
-                          >
-                            重置为默认位置
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {appSettingsPage === 'theme' && (
-                    <div className="max-expand-settings-section">
-                      <div className="settings-music-section">
-                        <div className="settings-music-label">主题模式</div>
-                        <div className="settings-music-hint">选择深色、浅色或跟随系统主题，切换后立即生效</div>
-                        <div className="settings-lyrics-source-options">
-                          {([
-                            { value: 'dark' as ThemeMode, label: '深色模式' },
-                            { value: 'light' as ThemeMode, label: '浅色模式' },
-                            { value: 'system' as ThemeMode, label: '跟随系统' },
-                          ]).map((opt) => (
-                            <button
-                              key={opt.value}
-                              className={`settings-lyrics-source-btn ${themeMode === opt.value ? 'active' : ''}`}
-                              type="button"
-                              onClick={() => {
-                                setThemeModeState(opt.value);
-                                applyThemeMode(opt.value).catch(() => {});
-                              }}
-                            >
-                              {opt.label}
-                            </button>
-                          ))}
-                        </div>
-
-                        <div className="settings-music-label" style={{ marginTop: 14 }}>灵动岛透明度</div>
-                        <div className="settings-music-hint">数值越低越透明（10% - 100%），调整后立即生效</div>
-                        <div className="settings-opacity-slider-row">
-                          <input
-                            className="settings-opacity-slider"
-                            type="range"
-                            min={10}
-                            max={100}
-                            step={1}
-                            value={islandOpacity}
-                            onChange={(e) => {
-                              const v = Number(e.target.value);
-                              const safe = Number.isFinite(v) ? Math.max(10, Math.min(100, Math.round(v))) : 100;
-                              setIslandOpacity(safe);
-                              applyIslandOpacity(safe);
-                              if (opacitySaveTimerRef.current) {
-                                clearTimeout(opacitySaveTimerRef.current);
-                              }
-                              opacitySaveTimerRef.current = setTimeout(() => {
-                                persistIslandOpacity(safe);
-                                opacitySaveTimerRef.current = null;
-                              }, 220);
-                            }}
-                            onBlur={() => {
-                              if (opacitySaveTimerRef.current) {
-                                clearTimeout(opacitySaveTimerRef.current);
-                                opacitySaveTimerRef.current = null;
-                              }
-                              persistIslandOpacity(islandOpacity);
-                            }}
-                          />
-                          <span className="settings-opacity-slider-value">{islandOpacity}%</span>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {appSettingsPage === 'behavior' && (
-                    <div className="max-expand-settings-section">
-                      <div className="settings-music-section">
-                        <div className="settings-music-label">鼠标移开自动收回 (重启后生效)</div>
-                        <div className="settings-music-hint">启用后，鼠标离开灵动岛时将自动回到空闲状态（若正在播放音乐则切到歌词态）</div>
-                        <div className="settings-hotkey-row" style={{ alignItems: 'center', marginTop: 8 }}>
-                          <label className="settings-music-hint" style={{ marginBottom: 0, display: 'flex', alignItems: 'center', gap: 8 }}>
-                            <input
-                              type="checkbox"
-                              checked={expandLeaveIdle}
-                              onChange={(e) => {
-                                setExpandLeaveIdle(e.target.checked);
-                                window.api.expandMouseleaveIdleSet(e.target.checked).catch(() => {});
-                              }}
-                            />
-                            展开态（Expand）鼠标移开后自动收回
-                          </label>
-                        </div>
-                        <div className="settings-hotkey-row" style={{ alignItems: 'center', marginTop: 6 }}>
-                          <label className="settings-music-hint" style={{ marginBottom: 0, display: 'flex', alignItems: 'center', gap: 8 }}>
-                            <input
-                              type="checkbox"
-                              checked={maxExpandLeaveIdle}
-                              onChange={(e) => {
-                                setMaxExpandLeaveIdle(e.target.checked);
-                                window.api.maxexpandMouseleaveIdleSet(e.target.checked).catch(() => {});
-                              }}
-                            />
-                            最大展开态（MaxExpand）鼠标移开后自动收回
-                          </label>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {appSettingsPage === 'autostart' && (
-                    <div className="max-expand-settings-section">
-                      <div className="settings-music-section">
-                        <div className="settings-music-label">实用工具</div>
-                        <div className="settings-music-hint">常用应用操作与日志工具</div>
-                        <div className="settings-hotkey-row" style={{ marginTop: 8, gap: 8 }}>
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={() => {
-                              window.api.quitApp();
-                            }}
-                          >
-                            关闭灵动岛
-                          </button>
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={() => {
-                              window.api.restartApp().catch(() => {});
-                            }}
-                          >
-                            重启灵动岛
-                          </button>
-                          <button
-                            className="settings-hotkey-btn"
-                            type="button"
-                            onClick={() => {
-                              window.api.openLogsFolder().catch(() => {});
-                            }}
-                          >
-                            打开日志文件夹
-                          </button>
-                        </div>
-
-                        <div className="settings-music-label" style={{ marginTop: 12 }}>开机自启</div>
-                        <div className="settings-music-hint">设置系统启动时是否自动运行灵动岛</div>
-                        <div className="settings-lyrics-source-options" style={{ marginTop: 8 }}>
-                          {([
-                            { value: 'disabled', label: '禁用' },
-                            { value: 'enabled', label: '启用' },
-                            { value: 'high-priority', label: '高优先级' },
-                          ] as const).map((opt) => (
-                            <button
-                              key={opt.value}
-                              className={`settings-lyrics-source-btn ${autostartMode === opt.value ? 'active' : ''}`}
-                              type="button"
-                              onClick={() => {
-                                setAutostartMode(opt.value);
-                                window.api.autostartSet(opt.value).catch(() => {});
-                              }}
-                            >
-                              {opt.label}
-                            </button>
-                          ))}
-                        </div>
-                        <div className="settings-music-hint" style={{ marginTop: 8 }}>
-                          {autostartMode === 'disabled' && '当前已禁用开机自启。'}
-                          {autostartMode === 'enabled' && '系统登录后将自动启动灵动岛。'}
-                          {autostartMode === 'high-priority' && '以高优先级启动，更早完成加载。'}
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-
-                <div className="settings-app-page-dots" aria-label="软件设置分页">
-                  {APP_SETTINGS_PAGES.map((page) => (
-                    <button
-                      key={page}
-                      className={`settings-app-page-dot ${appSettingsPage === page ? 'active' : ''}`}
-                      data-label={SETTINGS_TAB_LABELS[page]}
-                      type="button"
-                      onClick={() => setAppSettingsPage(page)}
-                      title={SETTINGS_TAB_LABELS[page]}
-                      aria-label={SETTINGS_TAB_LABELS[page]}
-                    />
-                  ))}
-                </div>
-              </div>
-
-            </div>
+            <AppSettingsSection
+              currentAppSettingsPageLabel={currentAppSettingsPageLabel}
+              appSettingsPage={appSettingsPage}
+              layoutConfig={layoutConfig}
+              OverviewPreviewComponent={OverviewPreview}
+              overviewWidgetOptions={OVERVIEW_WIDGET_OPTIONS}
+              updateLayout={updateLayout}
+              hideProcessFilter={hideProcessFilter}
+              setHideProcessFilter={setHideProcessFilter}
+              refreshRunningProcesses={refreshRunningProcesses}
+              hideProcessLoading={hideProcessLoading}
+              hideProcessList={hideProcessList}
+              toggleHideProcess={toggleHideProcess}
+              runningProcesses={runningProcesses}
+              hideProcessKeyword={hideProcessKeyword}
+              islandPositionOffset={islandPositionOffset}
+              applyIslandPositionOffset={applyIslandPositionOffset}
+              islandPositionInput={islandPositionInput}
+              setIslandPositionInput={setIslandPositionInput}
+              applyIslandPositionInput={applyIslandPositionInput}
+              islandPositionInputChanged={islandPositionInputChanged}
+              cancelIslandPositionInput={cancelIslandPositionInput}
+              themeMode={themeMode}
+              setThemeModeState={setThemeModeState}
+              applyThemeMode={applyThemeMode}
+              islandOpacity={islandOpacity}
+              applyIslandOpacity={applyIslandOpacity}
+              opacitySaveTimerRef={opacitySaveTimerRef}
+              setIslandOpacity={setIslandOpacity}
+              persistIslandOpacity={persistIslandOpacity}
+              expandLeaveIdle={expandLeaveIdle}
+              setExpandLeaveIdle={setExpandLeaveIdle}
+              maxExpandLeaveIdle={maxExpandLeaveIdle}
+              setMaxExpandLeaveIdle={setMaxExpandLeaveIdle}
+              autostartMode={autostartMode}
+              setAutostartMode={setAutostartMode}
+              appSettingsPages={APP_SETTINGS_PAGES}
+              settingsTabLabels={SETTINGS_TAB_LABELS}
+              setAppSettingsPage={setAppSettingsPage}
+            />
           )}
+
           {activeTab === 'network' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title">网络配置</div>
-              <div className="settings-music-section">
-                <div className="settings-music-label">请求超时时间</div>
-                <div className="settings-music-hint">设置网络请求的最长等待时间，网络较差时可适当增大</div>
-                <div className="settings-network-timeout-row">
-                  <div className="settings-lyrics-source-options">
-                    {NETWORK_TIMEOUT_OPTIONS.map((opt) => (
-                      <button
-                        key={opt.value}
-                        className={`settings-lyrics-source-btn ${networkTimeoutMs === opt.value ? 'active' : ''}`}
-                        type="button"
-                        onClick={() => {
-                          setNetworkTimeoutMs(opt.value);
-                          setCustomTimeoutInput(String(opt.value / 1000));
-                          saveNetworkConfig({ timeoutMs: opt.value });
-                        }}
-                      >
-                        {opt.label}
-                      </button>
-                    ))}
-                  </div>
-                  <div className={`settings-network-custom${NETWORK_TIMEOUT_OPTIONS.every(o => o.value !== networkTimeoutMs) ? ' active' : ''}`}>
-                    <input
-                      className="settings-network-custom-input"
-                      type="number"
-                      min="1"
-                      max="120"
-                      value={customTimeoutInput}
-                      onChange={(e) => setCustomTimeoutInput(e.target.value)}
-                      onBlur={() => {
-                        const sec = parseFloat(customTimeoutInput);
-                        if (!isNaN(sec) && sec >= 1) {
-                          const ms = Math.round(sec * 1000);
-                          setNetworkTimeoutMs(ms);
-                          saveNetworkConfig({ timeoutMs: ms });
-                        } else {
-                          setCustomTimeoutInput(String(networkTimeoutMs / 1000));
-                        }
-                      }}
-                      onKeyDown={(e) => {
-                        if (e.key === 'Enter') (e.target as HTMLInputElement).blur();
-                      }}
-                    />
-                    <span className="settings-network-custom-unit">秒</span>
-                  </div>
-                </div>
-              </div>
-
-            </div>
+            <NetworkSettingsSection
+              networkTimeoutMs={networkTimeoutMs}
+              customTimeoutInput={customTimeoutInput}
+              networkTimeoutOptions={NETWORK_TIMEOUT_OPTIONS}
+              setNetworkTimeoutMs={setNetworkTimeoutMs}
+              setCustomTimeoutInput={setCustomTimeoutInput}
+              saveNetworkConfig={saveNetworkConfig}
+            />
           )}
+
           {activeTab === 'weather' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title settings-app-title-line">
-                <span>天气配置</span>
-                <span className="settings-app-title-sub">- {currentWeatherSettingsPageLabel}</span>
-              </div>
-
-              <div className="settings-app-pages-layout settings-weather-pages-layout">
-                <div className="settings-app-page-main">
-                  {weatherSettingsPage === 'location' && (
-                    <div className="settings-music-section">
-                      <div className="settings-music-label">定位来源优先级</div>
-                      <div className="settings-music-hint">选择天气定位优先使用 IP 自动定位或自定义位置</div>
-                      <div className="settings-lyrics-source-options">
-                        {WEATHER_LOCATION_PRIORITY_OPTIONS.map((opt) => (
-                          <button
-                            key={opt.value}
-                            className={`settings-lyrics-source-btn ${weatherLocationPriority === opt.value ? 'active' : ''}`}
-                            type="button"
-                            onClick={() => {
-                              applyWeatherLocationPriority(opt.value).catch((error) => {
-                                setWeatherLocationConfigMessage({
-                                  type: 'error',
-                                  text: `切换优先级失败：${error instanceof Error ? error.message : '未知错误'}`,
-                                });
-                              });
-                            }}
-                          >
-                            {opt.label}
-                          </button>
-                        ))}
-                      </div>
-
-                      <div className="settings-hotkey-row">
-                        <label className="settings-field" style={{ flex: 1 }}>
-                          <span className="settings-field-label">城市名称</span>
-                          <input
-                            className="settings-field-input"
-                            type="text"
-                            placeholder="例如：杭州 / Tokyo / New York"
-                            value={weatherCustomCityInput}
-                            onChange={(e) => {
-                              setWeatherCustomCityInput(e.target.value);
-                            }}
-                          />
-                        </label>
-                      </div>
-
-                      <div className="settings-hotkey-row">
-                        <button
-                          className="settings-hotkey-btn"
-                          type="button"
-                          onClick={() => {
-                            testWeatherCustomLocation().catch((error) => {
-                              setWeatherCustomLocationTesting(false);
-                              setWeatherCustomLocationTestMessage({
-                                type: 'error',
-                                text: `测试失败：${error instanceof Error ? error.message : '未知错误'}`,
-                              });
-                            });
-                          }}
-                          disabled={weatherCustomLocationTesting}
-                        >
-                          {weatherCustomLocationTesting ? '测试中...' : '测试自定义位置（双接口）'}
-                        </button>
-                        <button
-                          className="settings-hotkey-btn"
-                          type="button"
-                          onClick={() => {
-                            saveWeatherLocationSettings().catch((error) => {
-                              setWeatherLocationConfigMessage({
-                                type: 'error',
-                                text: `保存失败：${error instanceof Error ? error.message : '未知错误'}`,
-                              });
-                            });
-                          }}
-                        >
-                          保存定位配置
-                        </button>
-                      </div>
-
-                      {weatherLocationConfigMessage && (
-                        <div className="settings-music-hint" style={{ color: weatherLocationConfigMessage.type === 'error' ? '#ff7f7f' : '#7be495' }}>
-                          {weatherLocationConfigMessage.text}
-                        </div>
-                      )}
-                      {weatherCustomLocationTestMessage && (
-                        <div className="settings-music-hint" style={{ color: weatherCustomLocationTestMessage.type === 'error' ? '#ff7f7f' : '#7be495' }}>
-                          {weatherCustomLocationTestMessage.text}
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  {weatherSettingsPage === 'provider' && (
-                    <div className="settings-music-section">
-                      <div className="settings-music-label">天气接口优先级</div>
-                      <div className="settings-music-hint">可选择优先使用 Open-Meteo 或 UAPI，失败时自动切换到另一源</div>
-                      <div className="settings-lyrics-source-options">
-                        {WEATHER_PROVIDER_OPTIONS.map((opt) => (
-                          <button
-                            key={opt.value}
-                            className={`settings-lyrics-source-btn ${weatherPrimaryProvider === opt.value ? 'active' : ''}`}
-                            type="button"
-                            onClick={() => {
-                              setWeatherPrimaryProvider(opt.value);
-                              saveWeatherProviderConfig({ primaryProvider: opt.value });
-                            }}
-                          >
-                            {opt.label}
-                          </button>
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                </div>
-
-                <div className="settings-app-page-dots" aria-label="天气配置分页">
-                  {WEATHER_SETTINGS_PAGES.map((page) => (
-                    <button
-                      key={page}
-                      className={`settings-app-page-dot ${weatherSettingsPage === page ? 'active' : ''}`}
-                      data-label={WEATHER_SETTINGS_PAGE_LABELS[page]}
-                      type="button"
-                      onClick={() => setWeatherSettingsPage(page)}
-                      title={WEATHER_SETTINGS_PAGE_LABELS[page]}
-                      aria-label={WEATHER_SETTINGS_PAGE_LABELS[page]}
-                    />
-                  ))}
-                </div>
-              </div>
-
-            </div>
+            <WeatherSettingsSection
+              currentWeatherSettingsPageLabel={currentWeatherSettingsPageLabel}
+              weatherSettingsPage={weatherSettingsPage}
+              weatherLocationPriorityOptions={WEATHER_LOCATION_PRIORITY_OPTIONS}
+              weatherLocationPriority={weatherLocationPriority}
+              applyWeatherLocationPriority={applyWeatherLocationPriority}
+              setWeatherLocationConfigMessage={setWeatherLocationConfigMessage}
+              weatherCustomCityInput={weatherCustomCityInput}
+              setWeatherCustomCityInput={setWeatherCustomCityInput}
+              testWeatherCustomLocation={testWeatherCustomLocation}
+              setWeatherCustomLocationTesting={setWeatherCustomLocationTesting}
+              setWeatherCustomLocationTestMessage={setWeatherCustomLocationTestMessage}
+              weatherCustomLocationTesting={weatherCustomLocationTesting}
+              saveWeatherLocationSettings={saveWeatherLocationSettings}
+              weatherLocationConfigMessage={weatherLocationConfigMessage}
+              weatherCustomLocationTestMessage={weatherCustomLocationTestMessage}
+              weatherProviderOptions={WEATHER_PROVIDER_OPTIONS}
+              weatherPrimaryProvider={weatherPrimaryProvider}
+              setWeatherPrimaryProvider={setWeatherPrimaryProvider}
+              saveWeatherProviderConfig={saveWeatherProviderConfig}
+              weatherSettingsPages={WEATHER_SETTINGS_PAGES}
+              weatherSettingsPageLabels={WEATHER_SETTINGS_PAGE_LABELS}
+              setWeatherSettingsPage={setWeatherSettingsPage}
+            />
           )}
+
           {activeTab === 'shortcut' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title">快捷键</div>
-              <div className="settings-hotkey-section">
-                <div className="settings-hotkey-label">隐藏/显示快捷键</div>
-                <div className="settings-hotkey-row">
-                  <input
-                    ref={hotkeyInputRef}
-                    className={`settings-hotkey-input ${hotkeyRecording ? 'recording' : ''}${hotkeyError ? ' error' : ''}`}
-                    type="text"
-                    readOnly
-                    value={hotkeyRecording ? '请按下快捷键组合…' : (hideHotkey || '未设置')}
-                    onFocus={() => {
-                      setHotkeyRecording(true);
-                      setHotkeyError('');
-                      window.api.hotkeySuspend().catch(() => {});
-                    }}
-                    onBlur={() => {
-                      setHotkeyRecording(false);
-                      window.api.hotkeyResume().catch(() => {});
-                    }}
-                    onKeyDown={handleHotkeyKeyDown}
-                  />
-                  <button
-                    className="settings-hotkey-btn"
-                    type="button"
-                    onClick={() => {
-                      setHotkeyRecording(true);
-                      hotkeyInputRef.current?.focus();
-                    }}
-                  >
-                    {hotkeyRecording ? '录入中' : '修改'}
-                  </button>
-                  {hideHotkey && (
-                    <button
-                      className="settings-hotkey-btn"
-                      type="button"
-                      onClick={() => {
-                        window.api.hotkeySet('').then((ok) => {
-                          if (ok) {
-                            setHideHotkey('');
-                            setHotkeyError('');
-                            setHotkeyRecording(false);
-                            hotkeyInputRef.current?.blur();
-                          }
-                        }).catch(() => {});
-                      }}
-                    >
-                      清除
-                    </button>
-                  )}
-                </div>
-                {hotkeyError && <div className="settings-hotkey-error">{hotkeyError}</div>}
-                <div className="settings-hotkey-hint">点击“修改”后按下组合键（如 Alt+X、Ctrl+Shift+H）</div>
-              </div>
-              <div className="settings-hotkey-section">
-                <div className="settings-hotkey-label">关闭灵动岛快捷键</div>
-                <div className="settings-hotkey-row">
-                  <input
-                    ref={quitHotkeyInputRef}
-                    className={`settings-hotkey-input ${quitHotkeyRecording ? 'recording' : ''}${quitHotkeyError ? ' error' : ''}`}
-                    type="text"
-                    readOnly
-                    value={quitHotkeyRecording ? '请按下快捷键组合…' : (quitHotkey || '未设置')}
-                    onFocus={() => {
-                      setQuitHotkeyRecording(true);
-                      setQuitHotkeyError('');
-                      window.api.hotkeySuspend().catch(() => {});
-                    }}
-                    onBlur={() => {
-                      setQuitHotkeyRecording(false);
-                      window.api.hotkeyResume().catch(() => {});
-                    }}
-                    onKeyDown={handleQuitHotkeyKeyDown}
-                  />
-                  <button
-                    className="settings-hotkey-btn"
-                    type="button"
-                    onClick={() => {
-                      setQuitHotkeyRecording(true);
-                      quitHotkeyInputRef.current?.focus();
-                    }}
-                  >
-                    {quitHotkeyRecording ? '录入中' : '修改'}
-                  </button>
-                  {quitHotkey && (
-                    <button
-                      className="settings-hotkey-btn"
-                      type="button"
-                      onClick={() => {
-                        window.api.quitHotkeySet('').then((ok) => {
-                          if (ok) {
-                            setQuitHotkey('');
-                            setQuitHotkeyError('');
-                            setQuitHotkeyRecording(false);
-                            quitHotkeyInputRef.current?.blur();
-                          }
-                        }).catch(() => {});
-                      }}
-                    >
-                      清除
-                    </button>
-                  )}
-                </div>
-                {quitHotkeyError && <div className="settings-hotkey-error">{quitHotkeyError}</div>}
-                <div className="settings-hotkey-hint">按下此快捷键将立即关闭灵动岛应用（如 Alt+Q、Ctrl+Shift+Q）</div>
-              </div>
-              <div className="settings-hotkey-section">
-                <div className="settings-hotkey-label">选区截图快捷键</div>
-                <div className="settings-hotkey-row">
-                  <input
-                    ref={screenshotHotkeyInputRef}
-                    className={`settings-hotkey-input ${screenshotHotkeyRecording ? 'recording' : ''}${screenshotHotkeyError ? ' error' : ''}`}
-                    type="text"
-                    readOnly
-                    value={screenshotHotkeyRecording ? '请按下快捷键组合…' : (screenshotHotkey || '未设置')}
-                    onFocus={() => {
-                      setScreenshotHotkeyRecording(true);
-                      setScreenshotHotkeyError('');
-                      window.api.hotkeySuspend().catch(() => {});
-                    }}
-                    onBlur={() => {
-                      setScreenshotHotkeyRecording(false);
-                      window.api.hotkeyResume().catch(() => {});
-                    }}
-                    onKeyDown={handleScreenshotHotkeyKeyDown}
-                  />
-                  <button
-                    className="settings-hotkey-btn"
-                    type="button"
-                    onClick={() => {
-                      setScreenshotHotkeyRecording(true);
-                      screenshotHotkeyInputRef.current?.focus();
-                    }}
-                  >
-                    {screenshotHotkeyRecording ? '录入中' : '修改'}
-                  </button>
-                  {screenshotHotkey && (
-                    <button
-                      className="settings-hotkey-btn"
-                      type="button"
-                      onClick={() => {
-                        window.api.screenshotHotkeySet('').then((ok) => {
-                          if (ok) {
-                            setScreenshotHotkey('');
-                            setScreenshotHotkeyError('');
-                            setScreenshotHotkeyRecording(false);
-                            screenshotHotkeyInputRef.current?.blur();
-                          }
-                        }).catch(() => {});
-                      }}
-                    >
-                      清除
-                    </button>
-                  )}
-                </div>
-                {screenshotHotkeyError && <div className="settings-hotkey-error">{screenshotHotkeyError}</div>}
-                <div className="settings-hotkey-hint">按下此快捷键将触发截图选区流程（如 Alt+A、Ctrl+Shift+A）</div>
-              </div>
-              <div className="settings-hotkey-section">
-                <div className="settings-hotkey-label">还原默认位置快捷键</div>
-                <div className="settings-hotkey-row">
-                  <input
-                    ref={resetPositionHotkeyInputRef}
-                    className={`settings-hotkey-input ${resetPositionHotkeyRecording ? 'recording' : ''}${resetPositionHotkeyError ? ' error' : ''}`}
-                    type="text"
-                    readOnly
-                    value={resetPositionHotkeyRecording ? '请按下快捷键组合…' : (resetPositionHotkey || '未设置')}
-                    onFocus={() => {
-                      setResetPositionHotkeyRecording(true);
-                      setResetPositionHotkeyError('');
-                      window.api.hotkeySuspend().catch(() => {});
-                    }}
-                    onBlur={() => {
-                      setResetPositionHotkeyRecording(false);
-                      window.api.hotkeyResume().catch(() => {});
-                    }}
-                    onKeyDown={handleResetPositionHotkeyKeyDown}
-                  />
-                  <button
-                    className="settings-hotkey-btn"
-                    type="button"
-                    onClick={() => {
-                      setResetPositionHotkeyRecording(true);
-                      resetPositionHotkeyInputRef.current?.focus();
-                    }}
-                  >
-                    {resetPositionHotkeyRecording ? '录入中' : '修改'}
-                  </button>
-                  {resetPositionHotkey && (
-                    <button
-                      className="settings-hotkey-btn"
-                      type="button"
-                      onClick={() => {
-                        window.api.resetPositionHotkeySet('').then((ok) => {
-                          if (ok) {
-                            setResetPositionHotkey('');
-                            setResetPositionHotkeyError('');
-                            setResetPositionHotkeyRecording(false);
-                            resetPositionHotkeyInputRef.current?.blur();
-                          }
-                        }).catch(() => {});
-                      }}
-                    >
-                      清除
-                    </button>
-                  )}
-                </div>
-                {resetPositionHotkeyError && <div className="settings-hotkey-error">{resetPositionHotkeyError}</div>}
-                <div className="settings-hotkey-hint">按下此快捷键将把灵动岛恢复到默认顶部居中位置</div>
-              </div>
-            </div>
+            <ShortcutSettingsSection
+              hotkeyInputRef={hotkeyInputRef}
+              hotkeyRecording={hotkeyRecording}
+              hotkeyError={hotkeyError}
+              hideHotkey={hideHotkey}
+              setHotkeyRecording={setHotkeyRecording}
+              setHotkeyError={setHotkeyError}
+              handleHotkeyKeyDown={handleHotkeyKeyDown}
+              setHideHotkey={setHideHotkey}
+              quitHotkeyInputRef={quitHotkeyInputRef}
+              quitHotkeyRecording={quitHotkeyRecording}
+              quitHotkeyError={quitHotkeyError}
+              quitHotkey={quitHotkey}
+              setQuitHotkeyRecording={setQuitHotkeyRecording}
+              setQuitHotkeyError={setQuitHotkeyError}
+              handleQuitHotkeyKeyDown={handleQuitHotkeyKeyDown}
+              setQuitHotkey={setQuitHotkey}
+              screenshotHotkeyInputRef={screenshotHotkeyInputRef}
+              screenshotHotkeyRecording={screenshotHotkeyRecording}
+              screenshotHotkeyError={screenshotHotkeyError}
+              screenshotHotkey={screenshotHotkey}
+              setScreenshotHotkeyRecording={setScreenshotHotkeyRecording}
+              setScreenshotHotkeyError={setScreenshotHotkeyError}
+              handleScreenshotHotkeyKeyDown={handleScreenshotHotkeyKeyDown}
+              setScreenshotHotkey={setScreenshotHotkey}
+              resetPositionHotkeyInputRef={resetPositionHotkeyInputRef}
+              resetPositionHotkeyRecording={resetPositionHotkeyRecording}
+              resetPositionHotkeyError={resetPositionHotkeyError}
+              resetPositionHotkey={resetPositionHotkey}
+              setResetPositionHotkeyRecording={setResetPositionHotkeyRecording}
+              setResetPositionHotkeyError={setResetPositionHotkeyError}
+              handleResetPositionHotkeyKeyDown={handleResetPositionHotkeyKeyDown}
+              setResetPositionHotkey={setResetPositionHotkey}
+            />
           )}
+
           {activeTab === 'music' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title settings-app-title-line">
-                <span>歌曲设置</span>
-                <span className="settings-app-title-sub">- {currentMusicSettingsPageLabel}</span>
-              </div>
-
-              <div className="settings-app-pages-layout settings-music-pages-layout">
-                <div className="settings-app-page-main">
-                  {musicSettingsPage === 'whitelist' && (
-                    <div className="settings-music-section">
-                      <div className="settings-music-label">播放器白名单</div>
-                      <div className="settings-music-hint">只有白名单内的播放器才会触发歌曲信息获取</div>
-                      <div className="settings-whitelist-list">
-                        {whitelist.map((item, idx) => (
-                          <div className="settings-whitelist-item" key={idx}>
-                            <span className="settings-whitelist-name">{item}</span>
-                            <button
-                              className="settings-whitelist-remove"
-                              type="button"
-                              title="移除"
-                              onClick={() => {
-                                const next = whitelist.filter((_, i) => i !== idx);
-                                setWhitelist(next);
-                                window.api.musicWhitelistSet(next).catch(() => {});
-                              }}
-                            >
-                              ×
-                            </button>
-                          </div>
-                        ))}
-                      </div>
-                      <div className="settings-whitelist-add-row">
-                        <input
-                          className={`settings-whitelist-input${whitelistInputError ? ' error' : ''}`}
-                          type="text"
-                          placeholder={whitelistInputError || '输入播放器进程名（如 Spotify.exe）'}
-                          value={whitelistDraft}
-                          onFocus={() => {
-                            if (whitelistInputError) setWhitelistInputError('');
-                          }}
-                          onChange={(e) => {
-                            setWhitelistDraft(e.target.value);
-                            if (whitelistInputError) setWhitelistInputError('');
-                          }}
-                          onKeyDown={(e) => {
-                            if (e.key === 'Enter') handleAddWhitelist();
-                          }}
-                        />
-                        <button
-                          className="settings-whitelist-add-btn"
-                          type="button"
-                          onClick={() => {
-                            handleAddWhitelist();
-                          }}
-                        >
-                          添加
-                        </button>
-                      </div>
-                      <div className="settings-whitelist-add-row" style={{ display: 'flex', alignItems: 'center' }}>
-                        <button
-                          className="settings-whitelist-add-btn"
-                          type="button"
-                          onClick={() => {
-                            if (whitelistInputError) setWhitelistInputError('');
-                            handleDetectSourceAppId().catch(() => {});
-                          }}
-                          disabled={detectingSourceAppId}
-                        >
-                          {detectingSourceAppId ? '获取中…' : '获取播放进程（测试）'}
-                        </button>
-                        {sourceAppDetectMessage && (
-                          <div
-                            className="settings-music-hint"
-                            style={{
-                              color: sourceAppDetectMessage.type === 'success' ? '#7df2a0' : '#ff8b8b',
-                              marginLeft: 10,
-                              marginBottom: 0,
-                              display: 'flex',
-                              alignItems: 'center',
-                            }}
-                          >
-                            {sourceAppDetectMessage.text}
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  )}
-
-                  {musicSettingsPage === 'lyrics' && (
-                    <div className="settings-music-section">
-                      <div className="settings-music-label">歌词源</div>
-                      <div className="settings-music-hint">自动模式根据 SMTC 检测到的播放器进程选择对应源，失败后依次尝试其他源，最后使用 LRCLIB 兜底</div>
-                      <div className="settings-lyrics-source-options">
-                        {LYRICS_SOURCE_OPTIONS.map((opt) => (
-                          <button
-                            key={opt.value}
-                            className={`settings-lyrics-source-btn ${lyricsSource === opt.value ? 'active' : ''}`}
-                            type="button"
-                            onClick={() => {
-                              setLyricsSource(opt.value);
-                              window.api.musicLyricsSourceSet(opt.value).catch(() => {});
-                            }}
-                          >
-                            {opt.label}
-                          </button>
-                        ))}
-                      </div>
-                      <div className="settings-music-label" style={{ marginTop: 12 }}>逐字扫光</div>
-                      <div className="settings-music-hint">启用后歌词将以逐字高亮方式显示</div>
-                      <div className="settings-hotkey-row" style={{ alignItems: 'center' }}>
-                        <label className="settings-music-hint" style={{ marginBottom: 0, display: 'flex', alignItems: 'center', gap: 8 }}>
-                          <input
-                            type="checkbox"
-                            checked={lyricsKaraoke}
-                            onChange={(e) => {
-                              setLyricsKaraoke(e.target.checked);
-                              window.api.musicLyricsKaraokeSet(e.target.checked).catch(() => {});
-                            }}
-                          />
-                          启用逐字扫光效果
-                        </label>
-                      </div>
-                      <div className="settings-music-label" style={{ marginTop: 12 }}>歌词时钟</div>
-                      <div className="settings-music-hint">在歌词界面封面与歌词之间显示当前北京时间</div>
-                      <div className="settings-hotkey-row" style={{ alignItems: 'center' }}>
-                        <label className="settings-music-hint" style={{ marginBottom: 0, display: 'flex', alignItems: 'center', gap: 8 }}>
-                          <input
-                            type="checkbox"
-                            checked={lyricsClock}
-                            onChange={(e) => {
-                              setLyricsClock(e.target.checked);
-                              window.api.musicLyricsClockSet(e.target.checked).catch(() => {});
-                            }}
-                          />
-                          显示当前时间
-                        </label>
-                      </div>
-                    </div>
-                  )}
-
-                  {musicSettingsPage === 'smtc' && (
-                    <div className="settings-music-section">
-                      <div className="settings-music-label">SMTC 自动取消订阅</div>
-                      <div className="settings-music-hint">用于清理长时间无更新的播放会话，默认永不取消订阅</div>
-                      <div className="settings-hotkey-row" style={{ alignItems: 'center' }}>
-                        <label className="settings-field" style={{ flex: 1 }}>
-                          <span className="settings-field-label">取消订阅时间（毫秒）</span>
-                          <input
-                            className="settings-field-input"
-                            type="number"
-                            min={1000}
-                            step={1000}
-                            value={musicSmtcUnsubscribeInput}
-                            disabled={musicSmtcNeverUnsubscribe}
-                            onChange={(e) => {
-                              setMusicSmtcUnsubscribeInput(e.target.value);
-                              if (musicSmtcConfigMessage) setMusicSmtcConfigMessage(null);
-                            }}
-                          />
-                        </label>
-                      </div>
-                      <div className="settings-hotkey-row" style={{ alignItems: 'center' }}>
-                        <label className="settings-music-hint" style={{ marginBottom: 0, display: 'flex', alignItems: 'center', gap: 8 }}>
-                          <input
-                            type="checkbox"
-                            checked={musicSmtcNeverUnsubscribe}
-                            onChange={(e) => {
-                              setMusicSmtcNeverUnsubscribe(e.target.checked);
-                              if (musicSmtcConfigMessage) setMusicSmtcConfigMessage(null);
-                            }}
-                          />
-                          永不取消订阅
-                        </label>
-                        <button
-                          className="settings-hotkey-btn"
-                          type="button"
-                          onClick={() => {
-                            saveMusicSmtcUnsubscribeConfig().catch((error) => {
-                              setMusicSmtcConfigMessage({
-                                type: 'error',
-                                text: `保存失败：${error instanceof Error ? error.message : '未知错误'}`,
-                              });
-                            });
-                          }}
-                        >
-                          保存
-                        </button>
-                      </div>
-                      {musicSmtcConfigMessage && (
-                        <div className="settings-music-hint" style={{ color: musicSmtcConfigMessage.type === 'error' ? '#ff8b8b' : '#7df2a0' }}>
-                          {musicSmtcConfigMessage.text}
-                        </div>
-                      )}
-                    </div>
-                  )}
-                </div>
-
-                <div className="settings-app-page-dots" aria-label="歌曲设置分页">
-                  {MUSIC_SETTINGS_PAGES.map((page) => (
-                    <button
-                      key={page}
-                      className={`settings-app-page-dot ${musicSettingsPage === page ? 'active' : ''}`}
-                      data-label={MUSIC_SETTINGS_PAGE_LABELS[page]}
-                      type="button"
-                      onClick={() => setMusicSettingsPage(page)}
-                      title={MUSIC_SETTINGS_PAGE_LABELS[page]}
-                      aria-label={MUSIC_SETTINGS_PAGE_LABELS[page]}
-                    />
-                  ))}
-                </div>
-              </div>
-            </div>
+            <MusicSettingsSection
+              currentMusicSettingsPageLabel={currentMusicSettingsPageLabel}
+              musicSettingsPage={musicSettingsPage}
+              whitelist={whitelist}
+              setWhitelist={setWhitelist}
+              whitelistInputError={whitelistInputError}
+              setWhitelistInputError={setWhitelistInputError}
+              whitelistDraft={whitelistDraft}
+              setWhitelistDraft={setWhitelistDraft}
+              handleAddWhitelist={handleAddWhitelist}
+              handleDetectSourceAppId={handleDetectSourceAppId}
+              detectingSourceAppId={detectingSourceAppId}
+              sourceAppDetectMessage={sourceAppDetectMessage}
+              lyricsSourceOptions={LYRICS_SOURCE_OPTIONS}
+              lyricsSource={lyricsSource}
+              setLyricsSource={setLyricsSource}
+              lyricsKaraoke={lyricsKaraoke}
+              setLyricsKaraoke={setLyricsKaraoke}
+              lyricsClock={lyricsClock}
+              setLyricsClock={setLyricsClock}
+              musicSmtcUnsubscribeInput={musicSmtcUnsubscribeInput}
+              setMusicSmtcUnsubscribeInput={setMusicSmtcUnsubscribeInput}
+              musicSmtcNeverUnsubscribe={musicSmtcNeverUnsubscribe}
+              setMusicSmtcNeverUnsubscribe={setMusicSmtcNeverUnsubscribe}
+              saveMusicSmtcUnsubscribeConfig={saveMusicSmtcUnsubscribeConfig}
+              setMusicSmtcConfigMessage={setMusicSmtcConfigMessage}
+              musicSmtcConfigMessage={musicSmtcConfigMessage}
+              musicSettingsPages={MUSIC_SETTINGS_PAGES}
+              musicSettingsPageLabels={MUSIC_SETTINGS_PAGE_LABELS}
+              setMusicSettingsPage={setMusicSettingsPage}
+            />
           )}
+
           {activeTab === 'ai' && (
-            <div className="max-expand-settings-section">
-              <div className="max-expand-settings-title">AI Agent</div>
-              <div className="settings-field-group">
-                <SettingsField
-                  label="API Key"
-                  value={aiConfig.apiKey}
-                  placeholder="sk-..."
-                  type="password"
-                  onChange={(v) => setAiConfig({ apiKey: v })}
-                />
-                <SettingsField
-                  label="API Endpoint"
-                  value={aiConfig.endpoint}
-                  placeholder="https://api.openai.com/v1"
-                  onChange={(v) => setAiConfig({ endpoint: v })}
-                />
-                <SettingsField
-                  label="模型"
-                  value={aiConfig.model}
-                  placeholder="gpt-4o-mini"
-                  onChange={(v) => setAiConfig({ model: v })}
-                />
-                <SettingsField
-                  label="MCP Endpoint"
-                  value={aiConfig.mcpEndpoint}
-                  placeholder="http://localhost:3000/mcp (可选)"
-                  onChange={(v) => setAiConfig({ mcpEndpoint: v })}
-                />
-                <div className="settings-field">
-                  <span className="settings-field-label">System Prompt</span>
-                  <div className="settings-prompt-area">
-                    {editingPrompt ? (
-                      <>
-                        <textarea
-                          ref={promptRef}
-                          className="settings-field-textarea"
-                          placeholder="你是一个有用的助手。"
-                          value={promptDraft}
-                          onChange={(e) => setPromptDraft(e.target.value)}
-                          onKeyDown={(e) => {
-                            if (e.key === 'Enter' && (e.ctrlKey || e.metaKey)) { e.preventDefault(); savePrompt(); }
-                          }}
-                          rows={3}
-                        />
-                        <button className="settings-prompt-btn save" onClick={savePrompt} type="button" title="保存 (Ctrl+Enter)">保存</button>
-                      </>
-                    ) : (
-                      <>
-                        <div className="settings-prompt-text">
-                          {aiConfig.systemPrompt || <span className="settings-prompt-empty">未设置</span>}
-                        </div>
-                        <button className="settings-prompt-btn edit" onClick={startEditPrompt} type="button" title="编辑 Prompt">编辑</button>
-                      </>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </div>
+            <AiSettingsSection
+              aiConfig={aiConfig}
+              editingPrompt={editingPrompt}
+              promptDraft={promptDraft}
+              promptRef={promptRef}
+              setAiConfig={setAiConfig}
+              setPromptDraft={setPromptDraft}
+              savePrompt={savePrompt}
+              startEditPrompt={startEditPrompt}
+              SettingsFieldComponent={SettingsField}
+            />
           )}
+
           {activeTab === 'update' && (
-            <div className="max-expand-settings-section settings-update">
-              <div className="max-expand-settings-title">更新设置</div>
-
-              <div className="settings-update-info-grid" style={{ display: 'flex', flexDirection: 'column', gap: 8, marginBottom: 0, fontSize: 12 }}>
-                <div style={{ display: 'flex', alignItems: 'center', gap: 16, flexWrap: 'wrap' }}>
-                  <span><span style={{ opacity: 0.6 }}>当前版本</span> <span style={{ fontWeight: 500 }}>eIsland v{aboutVersion || '…'}</span></span>
-                  <span style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-                    <span style={{ opacity: 0.6 }}>更新源</span>
-                    {UPDATE_SOURCES.map((s) => (
-                      <label key={s.key} style={{ display: 'inline-flex', alignItems: 'center', gap: 4, cursor: 'pointer', marginLeft: 4 }}>
-                        <input
-                          type="radio"
-                          name="update-source"
-                          value={s.key}
-                          checked={updateSource === s.key}
-                          onChange={() => setUpdateSource(s.key)}
-                          style={{ margin: 0 }}
-                        />
-                        <span>{s.label}</span>
-                      </label>
-                    ))}
-                  </span>
-                </div>
-                {(updateStatus === 'available' || updateStatus === 'downloading' || updateStatus === 'ready') && (
-                  <div style={{ display: 'flex', alignItems: 'center', gap: 8 }}>
-                    <span style={{ opacity: 0.6 }}>最新版本</span>
-                    <span style={{ fontWeight: 500, color: 'var(--accent-color, #4fc3f7)' }}>v{updateVersion}</span>
-                  </div>
-                )}
-              </div>
-
-              <div className="settings-about-update">
-                <div className="settings-about-update-row">
-                  {updateStatus === 'idle' && (
-                    <button className="settings-about-update-btn" type="button" onClick={handleCheckUpdate}>检查更新</button>
-                  )}
-                  {updateStatus === 'checking' && (
-                    <button className="settings-about-update-btn" type="button" disabled>检查中…</button>
-                  )}
-                  {updateStatus === 'latest' && (
-                    <button className="settings-about-update-btn" type="button" onClick={handleCheckUpdate}>已是最新版本</button>
-                  )}
-                  {updateStatus === 'available' && (
-                    <button className="settings-about-update-btn update-available" type="button" onClick={handleDownloadUpdate}>
-                      下载更新
-                    </button>
-                  )}
-                  {updateStatus === 'downloading' && (
-                    <div className="settings-about-update-progress">
-                      <div style={{ marginBottom: 4, fontSize: 12, opacity: 0.7 }}>正在从 {currentSourceLabel} 下载更新…</div>
-                      <div className="settings-about-update-progress-bar">
-                        <div
-                          className="settings-about-update-progress-fill"
-                          style={{ width: `${downloadProgress?.percent ?? 0}%` }}
-                        />
-                      </div>
-                      <span className="settings-about-update-progress-text">
-                        {downloadProgress ? `${Math.round(downloadProgress.percent)}% · ${(downloadProgress.bytesPerSecond / 1024 / 1024).toFixed(1)} MB/s` : '准备下载…'}
-                      </span>
-                    </div>
-                  )}
-                  {updateStatus === 'ready' && (
-                    <button className="settings-about-update-btn update-ready" type="button" onClick={handleInstallUpdate}>
-                      安装并重启
-                    </button>
-                  )}
-                  {updateStatus === 'error' && (
-                    <button className="settings-about-update-btn" type="button" onClick={handleCheckUpdate}>重试</button>
-                  )}
-                </div>
-                {updateStatus === 'error' && updateError && (
-                  <div className="settings-about-update-error" style={{ marginTop: 8, whiteSpace: 'pre-wrap' }}>{updateError.replace(/\\n/g, '\n')}</div>
-                )}
-              </div>
-            </div>
+            <UpdateSettingsSection
+              aboutVersion={aboutVersion}
+              updateSource={updateSource}
+              updateSources={UPDATE_SOURCES}
+              updateStatus={updateStatus}
+              updateVersion={updateVersion}
+              downloadProgress={downloadProgress}
+              currentSourceLabel={currentSourceLabel}
+              updateError={updateError}
+              onUpdateSourceChange={setUpdateSource}
+              onCheckUpdate={handleCheckUpdate}
+              onDownloadUpdate={handleDownloadUpdate}
+              onInstallUpdate={handleInstallUpdate}
+            />
           )}
-          {activeTab === 'about' && (
-            <div className="max-expand-settings-section settings-about">
-              <div className="max-expand-settings-title">关于软件</div>
 
-              {/* 作者信息 */}
-              <div className="settings-about-author">
-                <img className="settings-about-avatar" src={avatarImg} alt="作者头像" />
-                <div className="settings-about-author-info">
-                  <div className="settings-about-name">
-                    <a className="settings-about-github" href="https://github.com/JNTMTMTM" target="_blank" rel="noreferrer" title="GitHub 主页">
-                      <svg viewBox="0 0 16 16" width="14" height="14" fill="currentColor"><path d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27s1.36.09 2 .27c1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.01 8.01 0 0016 8c0-4.42-3.58-8-8-8z"/></svg>
-                    </a>
-                    鸡哥 <span className="settings-about-id">JNTMTMTM</span>
-                  </div>
-                  <div className="settings-about-version">eIsland v{aboutVersion}</div>
-                </div>
-              </div>
-
-              {/* 免费声明 */}
-              <div className="settings-about-notice">
-                本软件开源免费，如果你在任何地方付费购买了本软件，请立即退款并给差评。
-              </div>
-
-              {/* 链接 */}
-              <div className="settings-about-links">
-                <div className="settings-about-row">
-                  <span className="settings-about-label">官网</span>
-                  <a className="settings-about-link" href="https://www.pyisland.com" target="_blank" rel="noreferrer">www.pyisland.com</a>
-                </div>
-                <div className="settings-about-row">
-                  <span className="settings-about-label">文档站</span>
-                  <a className="settings-about-link" href="https://docs.pyisland.com" target="_blank" rel="noreferrer">docs.pyisland.com</a>
-                </div>
-                <div className="settings-about-row">
-                  <span className="settings-about-label">开源代码</span>
-                  <a className="settings-about-link" href="https://github.com/JNTMTMTM/eIsland" target="_blank" rel="noreferrer">github.com/JNTMTMTM/eIsland</a>
-                </div>
-                <div className="settings-about-row">
-                  <span className="settings-about-label">开源协议</span>
-                  <span className="settings-about-value">GPL-3.0</span>
-                </div>
-                <div className="settings-about-row">
-                  <span className="settings-about-label">图标库</span>
-                  <a className="settings-about-link" href="https://www.iconfont.cn/" target="_blank" rel="noreferrer">iconfont.cn</a>
-                </div>
-              </div>
-
-              {/* 开源依赖 */}
-              <div className="settings-about-deps">
-                <div className="settings-about-deps-title">开源框架 & 依赖</div>
-                <div className="settings-about-deps-grid">
-                  <span className="settings-about-dep">Electron</span>
-                  <span className="settings-about-dep">React</span>
-                  <span className="settings-about-dep">React DOM</span>
-                  <span className="settings-about-dep">TypeScript</span>
-                  <span className="settings-about-dep">Zustand</span>
-                  <span className="settings-about-dep">Tailwind CSS</span>
-                  <span className="settings-about-dep">Vite</span>
-                  <span className="settings-about-dep">electron-vite</span>
-                  <span className="settings-about-dep">electron-builder</span>
-                  <span className="settings-about-dep">react-markdown</span>
-                  <span className="settings-about-dep">react-datepicker</span>
-                  <span className="settings-about-dep">remark-gfm</span>
-                  <span className="settings-about-dep">@coooookies/windows-smtc-monitor</span>
-                  <span className="settings-about-dep">openmeteo</span>
-                  <span className="settings-about-dep">lunar-javascript</span>
-                  <span className="settings-about-dep">lyric-resolver</span>
-                  <span className="settings-about-dep">colorthief</span>
-                  <span className="settings-about-dep">lucide-react</span>
-                  <span className="settings-about-dep">@electron-toolkit/preload</span>
-                  <span className="settings-about-dep">@electron-toolkit/utils</span>
-                  <span className="settings-about-dep">@electron-toolkit/tsconfig</span>
-                  <span className="settings-about-dep">@tailwindcss/vite</span>
-                  <span className="settings-about-dep">@vitejs/plugin-react</span>
-                  <span className="settings-about-dep">PostCSS</span>
-                  <span className="settings-about-dep">Autoprefixer</span>
-                </div>
-              </div>
-
-              {/* 版权信息 */}
-              <div className="settings-about-footer">
-                <div className="settings-about-copyright">© JNTMTMTM, pyisland.com 版权所有</div>
-                <div className="settings-about-slogan">算法诠释一切 质疑即是认可</div>
-              </div>
-            </div>
-          )}
+          {activeTab === 'about' && <AboutSettingsSection aboutVersion={aboutVersion} />}
         </div>
       </div>
     </div>
